Move spec to typescript
